requirements.txt
# Core dependencies - pinned for stability
gradio>=5.0.0,<6.0.0
pydantic>=2.0.0,<3.0.0

# LangChain ecosystem
langchain>=0.3.0,<1.0.0
langchain-openai>=0.2.0,<1.0.0
langchain-core>=0.3.0,<1.0.0
langchain-community>=0.3.0,<1.0.0
langchain-experimental>=0.3.0,<1.0.0

# LangGraph
langgraph>=0.2.0,<1.0.0
langgraph-checkpoint-sqlite>=2.0.0,<3.0.0

# Database
aiosqlite>=0.20.0,<1.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0

# Browser automation
playwright>=1.40.0,<2.0.0

# Wikipedia
wikipedia>=1.4.0,<2.0.0
